third_party/blink/web_tests/external/wpt/wasm/core/global.wast.js-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] #17 Test that WebAssembly compilation fails
  assert_equals: expected true but got false
[FAIL] #18 Test that WebAssembly compilation fails
  assert_equals: expected true but got false
[FAIL] #144 Test that WebAssembly compilation fails
  assert_true: \n    at module (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:146:15)\n    at assert_invalid (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:174:3)\n    at global_wast_js (http://web-platform.test:8001/wasm/core/js/global.wast.js:232:1)\n    at http://web-platform.test:8001/wasm/core/js/global.wast.js:333:3 expected true got false
[FAIL] #145 Test that WebAssembly compilation fails
  assert_true: \n    at module (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:146:15)\n    at assert_invalid (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:174:3)\n    at global_wast_js (http://web-platform.test:8001/wasm/core/js/global.wast.js:235:1)\n    at http://web-platform.test:8001/wasm/core/js/global.wast.js:333:3 expected true got false
Harness: the test ran to completion.
